+++ system/event-script-engine/src/test/java/com/accenture/tasks/Greetings.java
@@ -39,11 +39,12 @@ public class Greetings implements TypedLambdaFunction<Map<String, Object>, Objec
     private static final String TIMEOUT = "timeout";
     private static final String CUSTOM = "custom";
     private static final String DEMO = "demo";
+    private static final String CONFLICT = "409";
     private static final String X_FLOW_ID = "x-flow-id";
 
     @Override
     public Object handleEvent(Map<String, String> headers, Map<String, Object> input, int instance)
-            throws InterruptedException {
+            throws InterruptedException, AppException {
         String exceptionTag = (String) input.get(EXCEPTION);
         if (exceptionTag != null) {
             if (TIMEOUT.equals(exceptionTag)) {
@@ -51,8 +52,12 @@ public class Greetings implements TypedLambdaFunction<Map<String, Object>, Objec
             } else if (CUSTOM.equals(exceptionTag)) {
                 return new EventEnvelope().setStatus(400).setBody(Map.of("error", "non-standard-format"));
             } else {
-                // just testing throwing an exception through a Mono reactive response
-                return Mono.create(emitter -> emitter.error(new AppException(403, "just a test")));
+                if (CONFLICT.equals(exceptionTag)) {
+                    throw new AppException(409, "Just a demo");
+                } else {
+                    // just testing throwing an exception through a Mono reactive response
+                    return Mono.create(emitter -> emitter.error(new AppException(403, "just a test")));
+                }
             }
         }
         if (input.containsKey(USER) && input.containsKey(GREETING)) {