+++ system/event-script-engine/src/test/java/com/accenture/tasks/HelloException.java
@@ -38,12 +38,14 @@ public class HelloException implements TypedLambdaFunction<Map<String, Object>, 
 
     @Override
     public Map<String, Object> handleEvent(Map<String, String> headers, Map<String, Object> input, int instance) {
-
         if (input.containsKey(STATUS) && input.containsKey(MESSAGE)) {
             Object stack = input.get("stack");
             if (stack instanceof String text) {
                 log.info("User defined exception handler got {}, rc={}, error={}, stack size={}",
                         headers, input.get(STATUS), input.get(MESSAGE), text.length());
+            }
+            if ("409".equals(String.valueOf(input.get(STATUS)))) {
+                throw new IllegalArgumentException("Demonstrate throwing exception at top level");
             }
             Map<String, Object> error = new HashMap<>();
             error.put(STATUS, input.get(STATUS));
@@ -54,5 +56,4 @@ public class HelloException implements TypedLambdaFunction<Map<String, Object>, 
             return Collections.emptyMap();
         }
     }
-
 }