+++ system/platform-core/src/main/java/org/platformlambda/core/util/Utility.java
@@ -653,11 +653,12 @@ public class Utility {
     }
 
     public boolean isPoJo(Object o) {
-        if (o == null || o instanceof Map || o instanceof String || o instanceof Number || o instanceof Boolean) {
+        if (o == null || o instanceof Map || o instanceof byte[] ||
+                o instanceof String || o instanceof Number || o instanceof Boolean) {
             return false;
         }
         String clsName = o.getClass().getName();
-        return !clsName.startsWith("java.");
+        return clsName.contains(".") && !clsName.startsWith("java.");
     }
 
     ///////////////////////////////////////